+++ src/bittytax/conv/parsers/__init__.py
@@ -57,6 +57,7 @@ from . import (
     ledgerlive,
     liquid,
     mercatox,
+    mexc,
     mymonero,
     nault,
     neonwallet,