+++ CHANGELOG.md
@@ -119,6 +119,7 @@
 - CoinTracking parser: added mapping for "Margin Profit", "Margin Loss" and "Margin Fee".
 - Koinly parser: added mapping for "Realized gain", "Loan", "Margin loan", "Margin fee", "Loan repayment" "Margin replayment", "Loan fee", "Other income", "Lending interest", "Cashback", "Salary" and "Fee refund".
 - Nexo parser: improved loan handling and added new loan transaction types.
+- Conversion tool: added parser for MEXC exchange.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))